file upload

+++ src/main/java/com/winter/app/regions/RegionDTO.java
@@ -6,14 +6,14 @@ public class RegionDTO {
 	//2. 모든 멤버변수의 게터세터 생성
 	//3. 생성자는 여러개 가능! 단, 기본생성자는 필수
 	//4. 멤버변수의 이름과 DATATYPE은 DB의 Table의 컬럼명과 DATATYPE 일치
-	private int region_id;
+	private Long region_id;
 	private String region_name;
 	
 	
-	public int getRegion_id() {
+	public Long getRegion_id() {
 		return region_id;
 	}
-	public void setRegion_id(int region_id) {
+	public void setRegion_id(Long region_id) {
 		this.region_id = region_id;
 	}
 	public String getRegion_name() {
